destroy core instance before next repaint
Currently, if you try to `destory()` the locomotive instance on useEffect
cleanup it throws an error:

`TypeError: Cannot read properties of undefined (reading 'destroy')``

It's coming from the core instance's destroy method, which is `undefined` by the
time the cleanup runs.

This commit fixes it by calling the `destroy()` method in the next repaint. Not sure
if this is the best fix but it Just Works™